move defaults to file
